requirements.txt
streamlit
qdrant-client
sentence-transformers
transformers
langdetect
groq
torch
pymupdf
rank_bm25
python-dotenv
nltk
certifi
